Add jsDocs, split into files, remove old files + minor refactoring
