+++ client/web/src/featureFlags/featureFlags.ts
@@ -24,7 +24,7 @@ class ProxyMap<K extends string, V extends boolean> extends Map<K, V> {
 
 // A union of all feature flags we currently have.
 // If there are no feature flags at the moment, this should be `never`.
-export type FeatureFlagName = 'open-beta-enabled' | 'quick-start-tour'
+export type FeatureFlagName = 'open-beta-enabled' | 'quick-start-tour-for-authenticated-users'
 
 export type FlagSet = ProxyMap<FeatureFlagName, boolean>
