--- client/web/src/gettingStartedTour/GettingStartedTour.module.scss
@@ -1,179 +1,0 @@
-:global(.theme-light) {
-    --getting-started-tour-oc-blue-bg: var(--oc-blue-1);
-    --getting-started-tour-oc-blue-border: var(--oc-blue-2);
-}
-
-:global(.theme-dark) {
-    --getting-started-tour-oc-blue-bg: var(--oc-blue-7);
-    --getting-started-tour-oc-blue-border: var(--oc-blue-8);
-}
-
-.card {
-    padding: 0.1875rem;
-    background: var(--marketing-gradient);
-    color: var(--body-color);
-    border-radius: 3px;
-
-    [data-reach-accordion-button] {
-        display: flex;
-        align-items: center;
-        margin-top: 0.5rem;
-        margin-bottom: 0.5rem;
-        cursor: pointer;
-
-        .arrow-icon-container {
-            margin-right: 0.25rem;
-            margin-left: -0.25rem;
-
-            svg {
-                transform: rotate(-90deg);
-            }
-        }
-
-        &[aria-expanded='true'] {
-            font-weight: bold;
-            .arrow-icon-container svg {
-                transform: none;
-            }
-        }
-    }
-
-    [data-reach-accordion-panel] {
-        margin-left: 1rem;
-        cursor: pointer;
-    }
-}
-
-.card-inner {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    background-color: var(--color-bg-1);
-    padding: 0.75rem;
-    height: 100%;
-}
-
-.card-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-}
-
-.card-title {
-    margin: 0;
-    font-size: 0.75rem;
-    font-weight: bold;
-    text-transform: uppercase;
-}
-
-.divider {
-    border: 0;
-    border-top: 1px solid var(--border-color);
-    width: 100%;
-    margin: 0.5rem 0;
-}
-
-.label {
-    margin-right: auto;
-    font-size: 0.875rem;
-    font-weight: normal;
-}
-
-.link {
-    &,
-    &:hover,
-    &:active {
-        text-decoration: none;
-        color: var(--body-color);
-    }
-}
-
-.is-fixed-height {
-    height: 7rem;
-    overflow-y: auto;
-    flex-grow: 1;
-
-    &::-webkit-scrollbar {
-        /* stylelint-disable-next-line declaration-property-unit-allowed-list */
-        margin-left: 4px;
-        /* stylelint-disable-next-line declaration-property-unit-allowed-list */
-        width: 4px;
-    }
-
-    &::-webkit-scrollbar-track {
-        border-radius: 3px;
-        box-shadow: inset 0 0 6px var(--color-bg-2);
-    }
-
-    &::-webkit-scrollbar-thumb {
-        border-radius: 3px;
-        box-shadow: inset 0 0 6px var(--text-muted);
-    }
-}
-
-.step {
-    display: flex;
-    margin-bottom: 0.5rem;
-}
-
-.progress-bar {
-    width: 0.9rem;
-    height: 0.9rem;
-    /* stylelint-disable-next-line declaration-property-unit-allowed-list */
-    margin-right: 1px;
-
-    // stylelint-disable-next-line selector-class-pattern
-    :global(.CircularProgressbar-trail) {
-        stroke: var(--dropdown-border-color);
-    }
-    // stylelint-disable-next-line selector-class-pattern
-    :global(.CircularProgressbar-path) {
-        stroke: var(--oc-green-4);
-    }
-}
-
-.footer-text {
-    margin-left: 0.25rem;
-}
-
-.text {
-    margin: 1rem 0;
-}
-
-.info-panel {
-    padding: 0.5rem;
-    border-radius: 3px;
-    background-color: var(--getting-started-tour-oc-blue-bg);
-    border: 1px solid var(--getting-started-tour-oc-blue-border);
-}
-
-.icon-muted {
-    color: var(--icon-color);
-    opacity: 0.4;
-}
-
-.info-icon {
-    color: var(--icon-color);
-    margin-right: 0.25rem;
-}
-
-.language-title {
-    font-size: 0.875rem;
-    text-transform: none;
-    font-weight: 500;
-    line-height: 1.25rem;
-}
-
-.language {
-    padding: 0.125rem 0.25rem;
-    &,
-    &:active,
-    &:focus,
-    &:hover {
-        background-color: var(--getting-started-tour-oc-blue-bg);
-    }
-
-    &:hover {
-        text-decoration: underline;
-    }
-}
