+++ client/web/src/search/panels/HomePanels.tsx
@@ -12,8 +12,8 @@ import { Tabs, Tab, TabList, TabPanel, TabPanels } from '@sourcegraph/wildcard'
 import { HomePanelsProps } from '..'
 import { AuthenticatedUser } from '../../auth'
 import { FeatureFlagProps } from '../../featureFlags/featureFlags'
-import { GettingStartedTour } from '../../gettingStartedTour/GettingStartedTour'
 import { CollaboratorsFragment, HomePanelsQueryResult, HomePanelsQueryVariables } from '../../graphql-operations'
+import { GettingStartedTour } from '../../tour/GettingStartedTour'
 import { eventLogger } from '../../tracking/eventLogger'
 
 import { CollaboratorsPanel } from './CollaboratorsPanel'
