+++ client/web/src/tour/components/Tour/types.ts
@@ -1,3 +1,6 @@
+/**
+ * Tour supported languages
+ */
 export enum TourLanguage {
     C = 'C',
     Go = 'Go',
@@ -8,10 +11,16 @@ export enum TourLanguage {
     Typescript = 'TypeScript',
 }
 
+/**
+ * Tour task
+ */
 export interface TourTaskType {
     title: string
     icon?: React.ReactNode
     steps: TourTaskStepType[]
+    /**
+     * Completion percentage, 0-100. Dynamically calculated field
+     */
     completed?: number
 }
 
@@ -22,8 +31,16 @@ export interface TourTaskStepType {
         type: 'video' | 'link' | 'restart'
         value: string | Record<TourLanguage, string>
     }
+    /**
+     * HTML string, which will be displayed in info box when navigating to a step link.
+     */
     info?: string
+    /**
+     * The step will be marked as completed only if one of the "completeAfterEvents" will be triggered
+     */
     completeAfterEvents?: string[]
-    // TODO: add jsDocs
+    /**
+     * Dynamically calculated field
+     */
     isCompleted?: boolean
 }
